ref(core): Restructure core module and fix CI/CD pipelines
Major architectural cleanup of the `core` application:
- Extracted DevOps scripts to `backend/scripts`.
- Moved shared utilities (throttling, signing, logging) to `backend/common`.
- Refactored settings into a modular `backend/settings` package.
- Cleaned up `core` to focus solely on `LandingPageSettings`.

CI/CD and Security fixes:
- Patched Log Injection vulnerability in `astrophotography` views.
- Fixed `flake8` linting errors, unused imports, and empty exception blocks.
- Resolved TypeScript errors in frontend tests.
- Verified all backend tests (148 passed) and formatted code with black/isort.
