+++ .pylintrc
@@ -2,7 +2,7 @@
 load-plugins = pylint_django
 
 [DJANGO]
-django-settings-module = backend.core.settings
+django-settings-module = backend.settings.base
 
 [MESSAGES CONTROL]
 disable = C0111
