+++ backend/core/__init__.py
@@ -3,11 +3,11 @@
 This will make sure the Celery app is always imported when Django starts
 so that shared_task will use this app.
 """
+
 try:
     from .celery import app as celery_app
 
     __all__ = ("celery_app",)
 except ImportError:
-    # This allows scripts like core.scripts:test to be imported on the host
     # where celery might not be installed (since tests run in Docker).
     pass
