+++ backend/translation/protocols.py
@@ -11,22 +11,22 @@ class TranslationAgentProtocol(Protocol):
         """
         Translates a place name with country context.
         """
-        ...
+        pass
 
     def translate_tag(self, text: str, target_lang_code: str) -> Optional[str]:
         """
         Translates a technical or descriptive tag.
         """
-        ...
+        pass
 
     def translate(self, text: str, target_lang_code: str) -> Optional[str]:
         """
         Translates plain text.
         """
-        ...
+        pass
 
     def translate_html(self, text: str, target_lang_code: str) -> Optional[str]:
         """
         Translates HTML content, preserving tags.
         """
-        ...
+        pass
