+++ backend/users/tests/test_views.py
@@ -3,7 +3,6 @@
 Tests for users views
 """
 
-
 import pytest
 from rest_framework import status
 from rest_framework.test import APIClient
